refactor: files structure
