+++ Exercises/quote-me/src/login/Login.tsx
@@ -1,10 +1,9 @@
 import React, { FunctionComponent, useState } from 'react'
 import { ConnectedUser } from '../contexts'
 import { User } from '../interfaces/user'
-import { LoginForm } from '../login-form/LoginForm'
+import { LoginForm } from './login-form/LoginForm'
 import { Logo } from '../logo/Logo'
-import HorizontalDiv from '../utils/HorizontalDiv'
-import './Login.css'
+import HorizontalDiv from '../utils/HorizontalDiv/HorizontalDiv'
 
 const Login: FunctionComponent = () => {
   const [connectedUser, setConnectedUser] = useState<User | undefined>(undefined)
