+++ Exercises/quote-me/src/login/login-form/LoginForm.css
@@ -2,8 +2,4 @@
     height: 1em;
     padding: 0.5em;
     color: red;
-<<<<<<< Updated upstream
 }
-=======
-}
->>>>>>> Stashed changes
